구현은 미궁속으로...

+++ Sandbox/EntryPoint.cpp
@@ -176,7 +176,7 @@ LRESULT CALLBACK WndProc(HWND hWnd, UINT message, WPARAM wParam, LPARAM lParam)
     switch (message)
     {
     case WM_SIZE:
-        EVENT_FIRE_DATA(Dive::eEventType::ChangedResolution, lParam);
+        EVENT_FIRE_DATA(Dive::eEventType::ChangedResolution, static_cast<unsigned int>(lParam));
         break;
     case WM_COMMAND:
         {
